+++ ml-core/src/main/scala/org/apache/spark/mllib/tree/impurity/Impurity.scala
@@ -81,7 +81,7 @@ private[spark] abstract class ImpurityAggregator(val statsSize: Int) extends Ser
    * @param allStats  Flat stats array, with stats for this (node, feature, bin) contiguous.
    * @param offset    Start index of stats for this (node, feature, bin).
    */
-  def update(allStats: Array[Double], offset: Int, label: Double, instanceWeight: Double): Unit
+  def update(allStats: Array[Double], offset: Int, label: Double, instanceWeight: Int): Unit
 
   /**
    * Get an [[ImpurityCalculator]] for a (node, feature, bin).
